+++ src/paypal-checkout/paypal-checkout.js
@@ -1580,6 +1580,12 @@ PayPalCheckout.prototype._attachPreloadPixel = function (options) {
   }
   if (options.merchantId) {
     preloadOptions["merchant-id"] = options.merchantId;
+  } else if (
+    this._configuration &&
+    this._configuration.gatewayConfiguration.merchantId
+  ) {
+    preloadOptions["merchant-id"] =
+      this._configuration.gatewayConfiguration.merchantId;
   }
 
   request = new XMLHttpRequest();
@@ -1612,6 +1618,13 @@ PayPalCheckout.prototype._formatPaymentResourceData = function (
     },
     shippingOptions: options.shippingOptions,
     payer_email: options.userAuthenticationEmail, // eslint-disable-line camelcase
+    source: constants.BT_SOURCE,
+    merchant: gatewayConfiguration.merchantId,
+
+    flowType:
+      options.flow === "vault"
+        ? constants.BT_FLOW_TYPES.VAULT
+        : constants.BT_FLOW_TYPES.EXPRESS_CHECKOUT,
   };
 
   if (options.hasOwnProperty("returnUrl")) {